Restructure context.rs in to context folder
